src/main/java/com/yong/service/MemberService.java
package com.yong.service;

import com.yong.domain.Member;
import com.yong.repository.MemberRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.dao.EmptyResultDataAccessException;
import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;

import java.util.List;

@Service
@Transactional
public class MemberService {
    @Autowired
    private MemberRepository memberRepository;

    public Long join(Member member){
        validateDuplicateMember(member);
        memberRepository.save(member);
        return member.getId();
    }

    private void validateDuplicateMember(Member member) {
        try{
            memberRepository.findByName(member.getName());
            throw new IllegalStateException("이미 존재하는 회원입니다. [" +member.getName()+ "]");
        }catch (EmptyResultDataAccessException e){

        }
    }

    public List<Member> findMembers(){
        return memberRepository.findAll();
    }

    public Member findOne(Long id){
        return memberRepository.findOne(id);
    }
}
